requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
Pillow==10.2.0
requests==2.31.0
python-magic-bin==0.4.14
